Fix issue in RTL where swipes on Launcher Smartspace are intercepted.
The `+ getScrollX()` translation is a duplicate, because
`mapCoordInSelfToDescendant` also does it internally. Same for the `+
getScrollY()`.

This wasn't an issue in LTR because the top left corner of the root view
is the same as the top left corner of the first page.  `getScrollX()`
returns 0 in that case.

In RTL, the second page is to the left of the first page. If the touch
is on the first page, `+ getScrollX()` translates it outside of the
first page. This incorrectly sets mIsEventOverFirstPagePinnedItem to
false, leading to the swipe being intercepted.

Bug: 240380590
Fix: 240380590
Test: manual
Change-Id: I51f534695401ce527da8d2158130a4d54b086f3d