Fix window x-axis movement after gesture ends.
Previously we setClamp to false right after the gesture ends.
So when the page settle animation gets the page offset, it would
get the unclamped version which would cause the window to move
in the x-axis movement even if the window didn't actually move
during the gesture.

Now we ensure that we setClamp to false after the page transition
ends so that the scroll offset is consistent all the way until
the end of the animation.

Bug: 258851206
Test: 1. swipe a bit diagonally to bring up taskbar with
         no x-axis window movement
      2. release
      3. there should be no x-axis movement on window

Change-Id: Ic9949d11f2e8bc7ccbd737555a3d5551f09456da