Set the ENABLE_DOWNLOAD_APP_UX_V2 flag to true
We will split the feature into two parts. The part that is guarded under ENABLE_DOWNLOAD_APP_UX_V2 is feature complete (everything except for the skeleton motion). The skeleton motion will be guarded by another flag (ENABLE_DOWNLOAD_APP_UX_V3). This is so that we have an improvement over the current version of restore icon treatment prior to release that we can start testing

Bug: 254858049
Test: Download an app and check how it looks when pending/donwloading/installing
Change-Id: I711f9d8acdb40689e71782acff6ef43907bb7435