Add EDU lottie animations for persistent and transient.
This CL introduces separate lottie assets for light and dark themes, as
well as each step. Each asset contains all of the steps currently but is
constrained to a min and max frame. In future CLs, we will optimize the
storage these take up by using the global assets and controlling min and
max frames programmatically, or trimming each individual asset.
Additionally, we should only be importing one theme variant and then
converting to the opposite programmatically, but the color mappings are
not ready yet.

Future polish to the EDU sheet, such as better handling the stashed
taskbar, will be introduced in future CLs.

Test: Manual in transient and persistent modes.
Bug: 217261955
Change-Id: I5b219ea02e25de0c5c887b5b5640f909e1219287