Fix bug with Taskbar second-app-selection animation
This patch makes it so that the correct animation will play when the user selects a second app from the Taskbar.

Previously, when the user selected a second app from the Taskbar, an animation would play showing the expansion of the app -- but it would come from the wrong place. The error occurred because when we calculate the animation start position, we assume that the starting view is located in the DragLayer, which fills the whole screen. But in cases where the starting view is located in the Taskbar, the root view is actually the TaskbarDragLayer, which only fills an area on the bottom part of the screen.

Fixed by adding a check to see if the starting view is part of the TaskbarActivityContext, and if so, adding an offset to calculate the correct DragLayer coordinates.

Fixes: 258692761
Test: Manual
Change-Id: I4026676fefe1e4c5cdf9d9a47d95c0d6e64672e8