Delay hiding Taskbar EDU until entering home or overview.
EDU will now end up hiding due to
TaskbarLauncherStateController#onStateChangeApplied closing all AFVs if
we are going to home or overview.

This is a quickfix to make it harder to accidentally close the EDU when
testing out the transient taskbar. We will eventually be moving to a
lighter EDU UI where having gestures occuring behind it will be less
awkward.

Test: Manual
Bug: 260769010
Change-Id: I477abe9339ed10cb3aed8d1a0fbe0e22ac14c7ab