Fix transition animations for launch-from-staged operation
This change makes it so that a smoother animation will play when the user cancels the splitscreen flow by tapping the staged app to launch the app in fullscreen. Since this was the only blocker to activating the feature, this change also enables launch-from-staged by default and removes the flag.

Fixes: 257513449
Test: Manual
Change-Id: Iabff2fc72bfcd6975d216b5cd86e0df7a772b308