Do not allow system gestures when dreaming.
Currently, TaskBar expansion is allowed when an activity context is
available. Since Dreams are implemented as an activity, this leads
to being able to expand the TaskBar over a dream. This changelist
updates the logic to not allow system gestures when dreaming.

Test: manual
Bug: 265755700
Change-Id: If58608c1e491da7906a2615681da0514fa49cfc8
Merged-In: If58608c1e491da7906a2615681da0514fa49cfc8