Load home properly even if AllSetActivity fails Go To Home Swipe.
If the user swipes to go home before the Launcher can startup properly,
the user will be stuck on a black screen after swiping up to go home
after reaching the AllSet page after completing the setup wizard. This
change forces the device to go home if the go to home swipe is going to
fail (aka if the Launcher hasn't finished initializing).

Bug: 261901912
Test: Manually verified the device doesn't get stuck like before. UX is
not perfect since the Launcher still needs to load for a few hundred
milliseconds, but the user doesn't get stuck on a black screen.

Change-Id: I8590e7ec1e341ba17db853900de7186284b07e93