Only interrupt taskbar state during gesture.
We only want to prevent updating taskbar stash state during
the swipe-to-overview gesture. Previously, we were blocking all
resumes into a "non-aligned-with-hotseat" state to prevent the
taskbar from hiding until the gesture completed. This, however,
also selected for situations where the launcher is resumed
directly into Overview outside of a transition (which is
triggered by some Tapl tests).

This narrows the check to only when resuming into a non-interactive
state (which basically just selects for BackgroundApp state).

Bug: 266246618
Test: TaplTestsQuickstep on a device with taskbar
Change-Id: I414ba7384f4ca3901211855b26c6c200319a2996