Improve resilience of testSplitFromOverview()
This patch makes it so that TaplTestsQuickstep#testSplitFromOverview() is less flaky.

Previously, this test worked by tapping the center of a Overview tile to confirm split select. However, sometimes the tile would be halfway offscreen, so it would accidentally tap an overlapping UI element (the staged first app) instead, causing the test to fail.

Fixed by using getCurrentTask() to always select an Overview tile that is fully onscreen.

Fixes: 267794149
Test: Manual on phone, confirmed test passing
Change-Id: I8b89509bb53a16e45aea4545562989e2e964de2d