Fix issue where bottom scrim appears in homescreen when it shouldn't
The bottom scrim should be shown for the Pixel Launcher, other third
party launchers show it when the 3 bottom navigation is activated.

The existing condition for 3 bottom navigation is no longer valid
because the insets at the bottom can be greather than 0 in other
situations other than 3 bottom navigation.

Test: the scrom should not show at all
Test: It appears in some situations where the taskbar is showing
and the scrim gets updates which not always happens so it's hard to
manually recreate, but there is a test that is flaky because of it
HomeScreenImageTabletTest#comparePixelTablet2023HomeScreen.
Fix: 267614579

Change-Id: Iacaf93011aa3f2f643186560fe34b4b26ce52314