Fix margins/padding in widget picker
Widget picker had weird issues because when we adjusted margin/padding on orientation change we were changing margin/padding of things that didn't have that amount set initially. The function onContentHorizontalMarginChanged was changing the margin/padding of recyclerviews and the stickyheaderlayout. This corrects the issue and there should be no margin/padding errors when we rotate, as the recyclerviews and stickyheaderlayout have the right margin set from the beginning.

Fix: 269348654
Test: Go to widget picker on tablet, rotate to landscape/portrait, rotate back to landscape/portrait. Verify that marins on the sides of widget picker are correct.
Change-Id: If5a26ae23b440c8a4a55cd0029042b2a5d5398b3