Widget picker closes and reopens when switching orientation on phones when it's not necessary
Widget picker closes and reopens when switching orientation on phones when it's not necessary. We should only close and reopen widget picker if it's tablet since that's when we want to inflate either the regular widget picker or the two pane widget picker. When we're on a phone, we always want to inflate the regular widget picker, so closing and re-inflating is not needed.

Fix: 271474746
Test: Change orientation in phones vs tablets. Verify that both still work.
Change-Id: I6062b056c4e7675bc830b6884761ec1f2eb20b1d