[Unfold animation] Restore clippings before clearing registered views
When launcher rotates during the fold/unfold animation
we clear views registered for the animation but we don't
restore the original clipping options that might be set.

This causes lost of the original property values
as the map that stores original values will store
updated values (after the animation preparation).

Bug: 273554726
Test: manual
Change-Id: I9c44a7ce4867ba0c3d6d1dca728d7c4cd3b378db