Fix animation canceling crash bug (short swipe during split select)
This patch fixes a bug where the user could cause a crash by making a gesture during the Overview > OverviewSplitSelect animation.

Within Overview, the user is able to make an upward gesture from the bottom of screen to return to Home. However, if the gesture is very short, the user doesn't go back to Home and instead stays in Overview. Under normal situations this doesn't cause any problems. But if we are in the middle of an animation, the short gesture actually triggers an animation cancel followed by an immediate goToState() to the same state that it was already in. This causes problems with the OverviewSplitSelect transition because reset() is called in the middle, clearing important split select data and causing a crash.

Fixed by changing a conditional to detect if we are this type of situation, and allowing the animation to play out in these cases.

Fixes: 272793237
Test: Manual
Change-Id: I4426204b9c8fc55853cf7df31a336ccaee2f5885