Add a way to fade out taskbar view when closing an app in the taskbar.
Prior to this change, if taskbar is showing and we are cloing an app
that is visible in the taskbar, the taskbar view would immediately
disappear.

Now we will fade out the view until the animation is 33% complete,
at which point the 'space' will be clear for the FloatingIconView
to settle at its final location.

Bug: 273961611
Test: open app in hotseat
      note that the taskbar view is gone immediately
      have taskbar visible, close app that is shown in taskbar
      note that the taskbar view fades out

Change-Id: I5589e2ce3033cfa19669d1bfaf568aef2a96015e