Stop enabling the Boot Aware Launcher prefs code during Tests after running LauncherPref unit tests.
There is something in the LauncherPrefs code that interacts with the
boot aware shared preference data that is not compatible with the
Widgets.Reorder tests. A todo has been added to fix that before
releasing that feature. For now, implementing a simple fix to stop
enabling this feature for while doing post submit testing.

Bug: 271434610
Test: Used go/abdt to verify that this CL fixes the post submit tests.
They no longer fail like they did before this was pushed.

Change-Id: Id6bd94f5d0e0ee9afc2ccea3a4808cf723a4e688