Fix bug with split screen not recognizing user differences
This patch fixes a bug where split screen did not fully support launching intents with different users.

The bug arose because SplitSelectStateController only had one place to store user information about the staged intent, mUser, but this disregarded the fact that the secondary app could also be passed in as an intent, and could belong to a different user from that of the initial app and the existing context. We need to support this case now since we now allow second-app selection from Taskbar.

Fixed by splitting the field into mInitialUser and mSecondUser, which will be tightly bound with mInitialTaskIntent and mSecondTaskIntent to make sure that Intents are always launched with the correct UserHandle.

Fixes: 275410160
Test: Manual
Change-Id: Ic904672769be8fd116180d457b36eb567c5ee304
Merged-In: Ic904672769be8fd116180d457b36eb567c5ee304
Merged-In: I1ec49c75d562e4309a41d98010f0eff113c81e9d